+++ RELEASE_NOTES.md
@@ -40,6 +40,10 @@ files in the same directory for embedding. Currently requires ECLiPSe, GNU
 Prolog, JIProlog, Quintus Prolog, SICStus Prolog, SWI-Prolog, XSB, or YAP
 as the backend compiler.
 
+* FIXED: Meta-predicate call optimization bug when a meta-argument is either
+a pre-compiled closure or a closure for a goal called in the context of the
+`user` pseudo-object.
+
 * FIXED: Pretty-printing of parametric object and parametric category
 identifiers in error and warning messages.